+++ collections/src/io/humb1t/orders/Action.java
@@ -1,4 +1,4 @@
-package io.humb1t;
+package io.humb1t.orders;
 
 @FunctionalInterface
 public interface Action {
