+++ collections/src/io/humb1t/orders/Order.java
@@ -1,4 +1,4 @@
-package io.humb1t;
+package io.humb1t.orders;
 
 import java.util.Objects;
