+++ collections/src/io/humb1t/orders/OrderStatus.java
@@ -1,4 +1,4 @@
-package io.humb1t;
+package io.humb1t.orders;
 
 public enum OrderStatus {
     NOT_STARTED, PROCESSING, COMPLETED
