+++ collections/src/io/humb1t/request/Request.java
@@ -1,4 +1,4 @@
-package io.humb1t;
+package io.humb1t.request;
 
 public class Request implements Runnable {
     private String data;
